backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
slowapi==0.1.8

# Email
jinja2==3.1.2

# Caching
redis==5.0.1

# ML/DL
torch==2.1.0
transformers==4.35.0
datasets==2.14.6
accelerate==0.24.1
peft==0.6.0
evaluate==0.4.0
rouge-score==0.1.2
sacrebleu==2.3.1
bert-score==0.3.13

# PDF Processing
pypdf2==3.0.1
pdfplumber==0.10.3
pytesseract==0.3.10
pillow==10.1.0
pdf2image==1.16.3

# Language Processing
langid==1.1.6
langdetect==1.0.9
indic-transliteration==1.8.1
regex==2023.10.3
sentencepiece==0.1.99
indic-nlp-library==0.81

# Data Processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2
nltk==3.8.1

# File Processing
python-magic==0.4.27
python-magic-bin==0.4.14

# Monitoring
prometheus-client==0.19.0
psutil==5.9.6

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
tzdata==2023.3
pyyaml==6.0.1
jsonschema==4.19.2

# Development
black==23.11.0
pytest==7.4.3
pytest-asyncio==0.21.1